"Refactored imports and file names in various Dart files, deleted desktop_layout.dart and order_details_screen.dart files."
