+++ lib/features/overview/views/OrdersListScreen.dart
@@ -3,7 +3,7 @@ import 'package:pos_dashboard_v1/core/utils/models/order_model.dart';
 import 'package:pdf/widgets.dart' as pw;
 import 'package:printing/printing.dart';
 import '../../../l10n/app_localizations.dart';
-import 'order_details_screen.dart';
+import 'OrderDetailsScreen.dart';
 
 class OrderListScreen extends StatelessWidget {
   final List<Order> orders;
